Update move card locations into their own folder
